pyproject.toml
[project]
name = "patchwork"
version = "0.1.0"
description = "A CLI agent for researching synth patches, controlling hardware synths via MIDI, and managing a patch library"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pydantic-ai",
    "rich",
    "python-dotenv",
    "python-rtmidi",
    "pyyaml",
]

[project.scripts]
patchwork = "patchwork.cli:main_cli"

[dependency-groups]
dev = ["pytest", "pytest-asyncio"]
